requirements.txt
Flask
Flask-CORS
pytesseract
pillow
openai
